// src/api/gen/types.gen.ts
export type PrayerName = 'fajr' | 'dhuhr' | 'asr' | 'maghrib' | 'isha' | 'jumah';
// ...
export type PrayerTimesOut = {
    /**
     * Fajr
     *
     * Time for Fajr prayer
     */
    fajr: string;
    /**
     * Shuruq
     *
     * Time for Shuruq
     */
    shuruq: string;
    /**
     * Dhuhr
     *
     * Time for Dhuhr prayer
     */
    dhuhr: string;
    /**
     * Asr
     *
     * Time for Asr prayer
     */
    asr: string;
    /**
     * Maghrib
     *
     * Time for Maghrib prayer
     */
    maghrib: string;
    /**
     * Isha
     *
     * Time for Isha prayer
     */
    isha: string;
    /**
     * Gregorian Date
     *
     * Gregorian date for this row
     */
    gregorian_date: string;
    /**
     * Hijri Date
     *
     * Hijri date for this row
     */
    hijri_date: string;
};